WIP, babel upgrade, coffeescript fails
