+++ .babelrc
@@ -1,7 +1,7 @@
 {
   "env": {
     "test": {
-      "presets": ["env"]
+      "presets": ["@babel/env"]
     }
   }
 }
