+++ .eslintrc
@@ -1,4 +1,9 @@
 {
+  "env": {
+    "node": true,
+    "jest": true,
+    "browser": true
+  },
   "root": true,
   "plugins": [
     "vue"
