+++ lib/compilers/babel-compiler.js
@@ -1,4 +1,4 @@
-const babel = require('babel-core')
+const babel = require('@babel/core')
 const loadBabelConfig = require('../load-babel-config.js')
 
 module.exports = function compileBabel (scriptContent, inputSourceMap, inlineConfig, vueJestConfig, filePath) {
@@ -19,7 +19,7 @@ module.exports = function compileBabel (scriptContent, inputSourceMap, inlineCon
   const babelOptions = Object.assign(sourceMapOptions, babelConfig)
   babelOptions.filename = 'unknown'
 
-  const res = babel.transform(scriptContent, babelOptions)
+  const res = babel.transformSync(scriptContent, babelOptions)
 
   return {
     code: res.code,
