+++ lib/compilers/coffee-compiler.js
@@ -6,6 +6,7 @@ module.exports = function (raw, vueJestConfig, filePath) {
   ensureRequire('coffee', ['coffeescript'])
   var coffee = require('coffeescript')
   var compiled
+
   try {
     compiled = coffee.compile(raw, {
       bare: true,
@@ -13,10 +14,12 @@ module.exports = function (raw, vueJestConfig, filePath) {
       transpile: loadBabelConfig(vueJestConfig, filePath)
     })
   } catch (err) {
-    throwError(err)
+    throw new Error(err)
+    // throwError(err)
   }
   return {
     code: compiled.js,
     map: compiled.v3SourceMap
   }
 }
+
