+++ lib/compilers/typescript-compiler.js
@@ -19,7 +19,7 @@ module.exports = function compileTypescript (scriptContent, vueJestConfig, fileP
   if (tsConfig.compilerOptions.module !== 'commonjs' && !loadBabelConfig(vueJestConfig, filePath)) {
     inlineBabelConfig = {
       plugins: [
-        require('babel-plugin-transform-es2015-modules-commonjs')
+        require('@babel/plugin-transform-modules-commonjs')
       ]
     }
   }
