+++ lib/process.js
@@ -38,7 +38,6 @@ module.exports = function (src, filePath, jestConfig) {
   if (parts.script && parts.script.src) {
     parts.script.content = fs.readFileSync(join(filePath, '..', parts.script.src), 'utf8')
   }
-
   const result = processScript(parts.script, vueJestConfig, filePath)
   const script = result.code
   const inputMap = result.sourceMap
