+++ test.sh
@@ -8,4 +8,4 @@ cp package.json temp-dir/package.json
 cp .babelrc temp-dir/.babelrc
 cp tsconfig.json temp-dir/tsconfig.json
 cd temp-dir
-yarn unit:run
+yarn unit:run -u
