--- test/.eslintrc
@@ -1,6 +1,0 @@
-{
-  "extends": '../.eslintrc',
-  "env": {
-    "jest": true
-  }
-}
